API Wrapper Decoderを別ファイルへ分離
